[INSTA-34029] Fix `Illegal option -o pipefail` on `dash`
The `pipefail` feature is independet from the age of the shell,
not even the latest version of `dash` would support this option.
And currently also on the latest `Ubuntu` images the `/bin/sh`
used by `instana-k8s-mustgather.sh` is symlinked to `dash`:

````
podman run --rm -it ubuntu:latest ls -lah /bin/sh
lrwxrwxrwx. 1 root root 4 Mar 31  2024 /bin/sh -> dash
````

Since `Ubuntu` and `dash` are not exactly uncommon,
and having to troubleshoot and edit the script in these cases
is a hassle, this commit checks in runtime for this option.

Signed-off-by: Ferenc Géczi <ferenc.geczi@ibm.com>